back/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.8.3
openai==1.54.3
